requirements-dev.txt
# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.11
requests-mock==1.11.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
boto3-stubs[lambda,dynamodb,secretsmanager]==1.34.14
aws-sam-cli==1.105.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Testing utilities
faker==20.1.0
freezegun==1.4.0
